got lots of it working, but still wip

+++ src/services/fluidType.tsx
@@ -1,14 +1,18 @@
-import { css, SerializedStyles } from '@emotion/react'
+import { SerializedStyles, css } from '@emotion/react'
 
-export const fluidType = (minVW: string, maxVW: string, minFontSize: string, maxFontSize: string): SerializedStyles => {
+export const fluidType = (
+  minVW: string,
+  maxVW: string,
+  minFontSize: string,
+  maxFontSize: string
+): SerializedStyles => {
   return css`
     font-size: ${minFontSize};
     @media (minwidth: ${minVW}) {
-      font-size:
-        calc(${minFontSize} +
-        ${parseInt(maxFontSize) - parseInt(minFontSize)} *
-        ((100vw - ${minVW}) /
-        ${parseInt(maxVW) - parseInt(minVW)}));
+      font-size: calc(
+        ${minFontSize} + ${parseFloat(maxFontSize) - parseFloat(minFontSize)} *
+          ((100vw - ${minVW}) / ${parseFloat(maxVW) - parseFloat(minVW)})
+      );
     }
     @media (min-width: ${maxVW}) {
       font-size: ${maxFontSize};
